Refactor HelperProvider into SmartDriverUtil, ElementUtil and DriverFactory
