+++ webfriends-selenium-elements/src/main/java/com/github/bordertech/webfriends/selenium/common/form/capability/SuggestableSelenium.java
@@ -2,9 +2,9 @@ package com.github.bordertech.webfriends.selenium.common.form.capability;
 
 import com.github.bordertech.webfriends.api.common.form.capability.Suggestable;
 import com.github.bordertech.webfriends.selenium.common.tag.SeleniumTags;
-import com.github.bordertech.webfriends.selenium.smart.driver.SmartHelper;
 import com.github.bordertech.webfriends.selenium.element.SElement;
 import com.github.bordertech.webfriends.selenium.element.form.select.SDatalist;
+import com.github.bordertech.webfriends.selenium.smart.driver.ElementUtil;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
@@ -28,7 +28,7 @@ public interface SuggestableSelenium extends Suggestable, SElement {
 		if (element == null) {
 			return null;
 		}
-		return SmartHelper.getProvider().wrapWebElement(getDriver(), element, SeleniumTags.DATALIST);
+		return ElementUtil.wrapWebElement(getDriver(), element, SeleniumTags.DATALIST);
 	}
 
 }
