+++ webfriends-selenium-elements/src/main/java/com/github/bordertech/webfriends/selenium/element/grouping/SUnorderedList.java
@@ -1,10 +1,11 @@
 package com.github.bordertech.webfriends.selenium.element.grouping;
 
-import com.github.bordertech.webfriends.selenium.common.tag.SeleniumTags;
-import com.github.bordertech.webfriends.selenium.element.AbstractSElement;
-import java.util.List;
 import com.github.bordertech.webfriends.api.element.grouping.HUnorderedList;
+import com.github.bordertech.webfriends.selenium.common.tag.SeleniumTags;
 import com.github.bordertech.webfriends.selenium.common.tags.STagUnorderedList;
+import com.github.bordertech.webfriends.selenium.element.AbstractSElement;
+import com.github.bordertech.webfriends.selenium.smart.driver.SmartDriverUtil;
+import java.util.List;
 
 /**
  * Selenium list (unordered) element.
@@ -21,7 +22,7 @@ public class SUnorderedList extends AbstractSElement implements HUnorderedList<S
 	@Override
 	public List<SUnorderedListItem> getListItems() {
 		if (items == null) {
-			items = getHelper().findWebFriends(getDriver(), getWebElement(), SeleniumTags.LISTITEM);
+			items = SmartDriverUtil.findWebFriends(getDriver(), getWebElement(), SeleniumTags.LISTITEM);
 		}
 		return items;
 	}
