+++ webfriends-selenium-elements/src/main/java/com/github/bordertech/webfriends/selenium/element/sections/SNav.java
@@ -6,6 +6,7 @@ import com.github.bordertech.webfriends.selenium.common.feature.ContainerWithChi
 import com.github.bordertech.webfriends.selenium.common.tag.SeleniumTags;
 import com.github.bordertech.webfriends.selenium.common.tags.STagNav;
 import com.github.bordertech.webfriends.selenium.element.AbstractSElement;
+import com.github.bordertech.webfriends.selenium.smart.driver.SmartDriverUtil;
 
 /**
  * Selenium nav element.
@@ -20,7 +21,7 @@ public class SNav extends AbstractSElement implements HNav,
 
 	@Override
 	public String getSectionTitle() {
-		HeadingElementSelenium heading = getHelper().findHeading(getDriver(), getWebElement());
+		HeadingElementSelenium heading = SmartDriverUtil.findHeading(getDriver(), getWebElement());
 		return heading == null ? null : heading.getHeadingText();
 	}
