+++ webfriends-selenium-elements/src/main/java/com/github/bordertech/webfriends/selenium/smart/driver/SmartDriver.java
@@ -6,13 +6,11 @@ import com.github.bordertech.webfriends.selenium.common.tag.TagTypeSelenium;
 import com.github.bordertech.webfriends.selenium.element.SElement;
 import com.github.bordertech.webfriends.selenium.element.form.SButton;
 import com.github.bordertech.webfriends.selenium.util.driver.FriendDriver;
+import java.util.List;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
-import org.openqa.selenium.SearchContext;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-
-import java.util.List;
 
 /**
  * Smart Elements Web Driver.
@@ -23,6 +21,7 @@ public class SmartDriver extends FriendDriver {
 	 * Default constructor.
 	 */
 	public SmartDriver() {
+		// Do nothing
 	}
 
 	/**
@@ -30,6 +29,15 @@ public class SmartDriver extends FriendDriver {
 	 */
 	public SmartDriver(final WebDriver webDriver) {
 		super(webDriver);
+	}
+
+	/**
+	 * Set focus to this element.
+	 *
+	 * @param element the element to focus
+	 */
+	public void setFocus(final SElement element) {
+		setFocus(element.getWebElement());
 	}
 
 	/**
@@ -42,30 +50,6 @@ public class SmartDriver extends FriendDriver {
 		return findButton(text, false);
 	}
 
-
-	/**
-	 * Find a button with the matching text.
-	 *
-	 * @param text          the text of the button
-	 * @param searchContext the search context to use
-	 * @return the button element or null
-	 */
-	public SButton findButton(final String text, final SearchContext searchContext) {
-		return getHelper().findButton(this, searchContext, text, false);
-	}
-
-	/**
-	 * Find a button with the matching text with search context.
-	 *
-	 * @param text          the text of the button
-	 * @param searchContext the search context to use
-	 * @param partial       true if partial match
-	 * @return the button element or null
-	 */
-	public SButton findButton(final String text, final SearchContext searchContext, final boolean partial) {
-		return getHelper().findButton(this, searchContext, text, partial);
-	}
-
 	/**
 	 * Find a button with the matching text.
 	 *
@@ -74,7 +58,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the button element or null
 	 */
 	public SButton findButton(final String text, final boolean partial) {
-		return getHelper().findButton(this, getWebDriver(), text, partial);
+		return SmartDriverUtil.findButton(this, getWebDriver(), text, partial);
 	}
 
 	/**
@@ -90,33 +74,6 @@ public class SmartDriver extends FriendDriver {
 	}
 
 	/**
-	 * Find a labeled element with the exact text with search context.
-	 *
-	 * @param <T>           the element tag type
-	 * @param elementTag    the element tag
-	 * @param searchContext the search context
-	 * @param text          the text of the button
-	 * @return the labeled element or null
-	 */
-	public <T extends LabelableSelenium> T findLabeled(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final String text) {
-		return findLabeled(elementTag, searchContext, text, false);
-	}
-
-	/**
-	 * Find a labeled element with the matching text.
-	 *
-	 * @param <T>           the element tag type
-	 * @param elementTag    the element tag
-	 * @param searchContext the search context
-	 * @param text          the text of the button
-	 * @param partial       true if partial match
-	 * @return the labeled element or null
-	 */
-	public <T extends LabelableSelenium> T findLabeled(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final String text, final boolean partial) {
-		return getHelper().findLabeled(this, searchContext, elementTag, text, partial);
-	}
-
-	/**
 	 * Find a labeled element with the matching text.
 	 *
 	 * @param <T> the element tag type
@@ -126,7 +83,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the labeled element or null
 	 */
 	public <T extends LabelableSelenium> T findLabeled(final TagTypeSelenium<T> elementTag, final String text, final boolean partial) {
-		return getHelper().findLabeled(this, getWebDriver(), elementTag, text, partial);
+		return SmartDriverUtil.findLabeled(this, getWebDriver(), elementTag, text, partial);
 	}
 
 	/**
@@ -141,33 +98,6 @@ public class SmartDriver extends FriendDriver {
 		return findWithText(elementTag, text, true);
 	}
 
-
-	/**
-	 * Find the first element with the tag type and matching text with search context.
-	 *
-	 * @param <T>        the element tag type
-	 * @param elementTag the element tag
-	 * @param searchContext the search context
-	 * @param text       the text of the button
-	 * @return the labeled element or null
-	 */
-	public <T extends SElement> T findWithText(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final String text) {
-		return getHelper().findWithText(this, searchContext, elementTag, text, true);
-	}
-
-	/**
-	 * Find the first element with the tag type and matching text.
-	 *
-	 * @param <T>        the element tag type
-	 * @param elementTag the element tag
-	 * @param text       the text of the button
-	 * @param partial    true if partial match
-	 * @return the labeled element or null
-	 */
-	public <T extends SElement> T findWithText(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final String text, final boolean partial) {
-		return getHelper().findWithText(this, searchContext, elementTag, text, partial);
-	}
-
 	/**
 	 * Find the first element with the tag type and matching text.
 	 *
@@ -178,7 +108,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the labeled element or null
 	 */
 	public <T extends SElement> T findWithText(final TagTypeSelenium<T> elementTag, final String text, final boolean partial) {
-		return getHelper().findWithText(this, getWebDriver(), elementTag, text, partial);
+		return SmartDriverUtil.findWithText(this, getWebDriver(), elementTag, text, partial);
 	}
 
 	/**
@@ -189,33 +119,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the web friend wrapper for the matching element, or null if no match
 	 */
 	public <T extends SElement> T findWebFriend(final TagTypeSelenium<T> elementTag) {
-		return getHelper().findWebFriend(this, getWebDriver(), elementTag);
-	}
-
-	/**
-	 * Find the first child element with this tag type.
-	 *
-	 * @param <T> the web friends element type
-	 * @param elementTag the element tag
-	 * @param searchContext the search context to use
-	 * @return the web friend wrapper for the matching element, or null if no match
-	 */
-	public <T extends SElement> T findWebFriend(final TagTypeSelenium<T> elementTag, final SearchContext searchContext) {
-		return getHelper().findWebFriend(this, searchContext, elementTag);
-	}
-
-
-	/**
-	 * Find the first child element with this Tag Type and By.
-	 *
-	 * @param <T>           the web friends element type
-	 * @param searchContext the search context to use
-	 * @param elementTag    the element tag
-	 * @param by            the by condition
-	 * @return the web friend wrapper for the matching element, or null if no match
-	 */
-	public <T extends SElement> T findWebFriend(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final By by) {
-		return getHelper().findWebFriend(this, searchContext, elementTag, by);
+		return SmartDriverUtil.findWebFriend(this, getWebDriver(), elementTag);
 	}
 
 	/**
@@ -227,20 +131,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the web friend wrapper for the matching element, or null if no match
 	 */
 	public <T extends SElement> T findWebFriend(final TagTypeSelenium<T> elementTag, final By by) {
-		return findWebFriend(elementTag, getWebDriver(), by);
-	}
-
-	/**
-	 * Find the nth (starting at 1) child element with this tag type with search context.
-	 *
-	 * @param <T>        the web friends element type
-	 * @param elementTag the element tag
-	 * @param searchContext the search context to use
-	 * @param idx        the nth element (starting at 1)
-	 * @return the web friend wrapper for the matching element, or null if no match
-	 */
-	public <T extends SElement> T findWebFriendNth(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final int idx) {
-		return getHelper().findWebFriendNth(this, searchContext, elementTag, idx);
+		return SmartDriverUtil.findWebFriend(this, getWebDriver(), elementTag, by);
 	}
 
 	/**
@@ -252,21 +143,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the web friend wrapper for the matching element, or null if no match
 	 */
 	public <T extends SElement> T findWebFriendNth(final TagTypeSelenium<T> elementTag, final int idx) {
-		return getHelper().findWebFriendNth(this, getWebDriver(), elementTag, idx);
-	}
-
-	/**
-	 * Find the nth (starting at 1) child element with this tag type and By with search context.
-	 *
-	 * @param <T>        the web friends element type
-	 * @param elementTag the element tag
-	 * @param searchContext the search context to use
-	 * @param by         the by condition
-	 * @param idx        the nth element (starting at 1)
-	 * @return the web friend wrapper for the matching element, or null if no match
-	 */
-	public <T extends SElement> T findWebFriendNth(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final By by, final int idx) {
-		return getHelper().findWebFriendNth(this, searchContext, elementTag, by, idx);
+		return SmartDriverUtil.findWebFriendNth(this, getWebDriver(), elementTag, idx);
 	}
 
 	/**
@@ -279,7 +156,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the web friend wrapper for the matching element, or null if no match
 	 */
 	public <T extends SElement> T findWebFriendNth(final TagTypeSelenium<T> elementTag, final By by, final int idx) {
-		return this.findWebFriendNth(elementTag, getWebDriver(), by, idx);
+		return SmartDriverUtil.findWebFriendNth(this, getWebDriver(), elementTag, by, idx);
 	}
 
 	/**
@@ -290,7 +167,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the web friend wrapper for the matching element, or null if no match
 	 */
 	public <T extends SElement> List<T> findWebFriends(final TagTypeSelenium<T> elementTag) {
-		return getHelper().findWebFriends(this, getWebDriver(), elementTag);
+		return SmartDriverUtil.findWebFriends(this, getWebDriver(), elementTag);
 	}
 
 	/**
@@ -302,32 +179,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the web friend wrapper for the matching element, or null if no match
 	 */
 	public <T extends SElement> List<T> findWebFriends(final TagTypeSelenium<T> elementTag, final By by) {
-		return findWebFriends(elementTag, getWebDriver(), by);
-	}
-
-	/**
-	 * Find the child elements with this tag type and By.
-	 *
-	 * @param <T> the web friends element type
-	 * @param elementTag the element tag
-	 * @param searchContext the search context to use
-	 * @param by the BY condition
-	 * @return the web friend wrapper for the matching element, or null if no match
-	 */
-	public <T extends SElement> List<T> findWebFriends(final TagTypeSelenium<T> elementTag, final SearchContext searchContext, final By by) {
-		return getHelper().findWebFriends(this, searchContext, elementTag, by);
-	}
-
-	/**
-	 * Find the child elements with this tag type and By.
-	 *
-	 * @param <T> the web friends element type
-	 * @param elementTag the element tag
-	 * @param searchContext the search context to use
-	 * @return the web friend wrapper for the matching element, or null if no match
-	 */
-	public <T extends SElement> List<T> findWebFriends(final TagTypeSelenium<T> elementTag, final SearchContext searchContext) {
-		return getHelper().findWebFriends(this, searchContext, elementTag);
+		return SmartDriverUtil.findWebFriends(this, getWebDriver(), elementTag, by);
 	}
 
 	/**
@@ -337,18 +189,18 @@ public class SmartDriver extends FriendDriver {
 	 * @return the matching element or null
 	 */
 	public WebElement findWebElement(final TagTypeSelenium elementTag) {
-		return getHelper().findWebElement(getWebDriver(), elementTag);
+		return SmartDriverUtil.findWebElement(getWebDriver(), elementTag);
 	}
 
 	/**
 	 * Find the first child element for a tag type.
 	 *
 	 * @param elementTag the element tag
-	 * @param relative   make the context relative
+	 * @param relative make the context relative
 	 * @return the matching element or null
 	 */
 	public WebElement findWebElement(final TagTypeSelenium elementTag, final boolean relative) {
-		return getHelper().findWebElement(getWebDriver(), elementTag, relative);
+		return SmartDriverUtil.findWebElement(getWebDriver(), elementTag, relative);
 	}
 
 	/**
@@ -359,7 +211,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the matching element or null
 	 */
 	public WebElement findWebElement(final TagTypeSelenium elementTag, final By by) {
-		return getHelper().findWebElement(getWebDriver(), elementTag, by);
+		return SmartDriverUtil.findWebElement(getWebDriver(), elementTag, by);
 	}
 
 	/**
@@ -370,7 +222,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the matching element or null
 	 */
 	public WebElement findWebElementNth(final TagTypeSelenium elementTag, final int idx) {
-		return getHelper().findWebElementNth(getWebDriver(), elementTag, idx);
+		return SmartDriverUtil.findWebElementNth(getWebDriver(), elementTag, idx);
 	}
 
 	/**
@@ -382,7 +234,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the matching element or null
 	 */
 	public WebElement findWebElementNth(final TagTypeSelenium elementTag, final By by, final int idx) {
-		return getHelper().findWebElementNth(getWebDriver(), elementTag, by, idx);
+		return SmartDriverUtil.findWebElementNth(getWebDriver(), elementTag, by, idx);
 	}
 
 	/**
@@ -392,7 +244,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return the matching elements.
 	 */
 	public List<WebElement> findWebElements(final TagTypeSelenium elementTag) {
-		return getHelper().findWebElements(getWebDriver(), elementTag);
+		return SmartDriverUtil.findWebElements(getWebDriver(), elementTag);
 	}
 
 	/**
@@ -403,24 +255,26 @@ public class SmartDriver extends FriendDriver {
 	 * @return the matching elements.
 	 */
 	public List<WebElement> findWebElements(final TagTypeSelenium elementTag, final By by) {
-		return getHelper().findWebElements(getWebDriver(), elementTag, by);
+		return SmartDriverUtil.findWebElements(getWebDriver(), elementTag, by);
 	}
 
 	/**
 	 *
 	 * Finds the webfriend of a child elements parent
+	 *
 	 * @param childWebElement the child web element
 	 * @param elementTag the element tag to wrap element with
 	 * @param <T> the web friends element type returned
 	 * @return
 	 */
 	public <T extends SElement> T findWebFriendParent(final WebElement childWebElement, final TagTypeSelenium<T> elementTag) {
-		return getHelper().findWebFriendParent(this, childWebElement, elementTag);
+		return SmartDriverUtil.findWebFriendParent(this, childWebElement, elementTag);
 	}
 
 	/**
 	 *
 	 * Finds the webfriend of a child elements parent
+	 *
 	 * @param childElement the child element
 	 * @param elementTag the element tag to wrap element with
 	 * @param <T> the web friends element type returned
@@ -428,7 +282,7 @@ public class SmartDriver extends FriendDriver {
 	 * @return
 	 */
 	public <T extends SElement, I extends SElement> T findWebFriendParent(final I childElement, final TagTypeSelenium<T> elementTag) {
-		return getHelper().findWebFriendParent(this, childElement, elementTag);
+		return SmartDriverUtil.findWebFriendParent(this, childElement, elementTag);
 	}
 
 	/**
@@ -456,14 +310,6 @@ public class SmartDriver extends FriendDriver {
 		keys.append(Keys.SHIFT);
 		keys.append(Keys.TAB);
 		body.sendKeys(Keys.chord(keys.toString()));
-	}
-
-	/**
-	 * @return the helper
-	 */
-	@Override
-	public SmartHelperProvider getHelper() {
-		return SmartHelper.getProvider();
 	}
 
 }
