--- webfriends-selenium-elements/src/main/java/com/github/bordertech/webfriends/selenium/smart/driver/SmartHelper.java
@@ -1,24 +1,0 @@
-package com.github.bordertech.webfriends.selenium.smart.driver;
-
-import com.github.bordertech.webfriends.selenium.util.driver.Helper;
-
-/**
- * Recasts the helper to the Smart Helper.
- */
-public final class SmartHelper {
-
-	/**
-	 * @return the smart elements helper
-	 */
-	public static SmartHelperProvider getProvider() {
-		return (SmartHelperProvider) Helper.getProvider();
-	}
-
-	/**
-	 * Private constructor.
-	 */
-	private SmartHelper() {
-		// No IMPL
-	}
-
-}
