+++ webfriends-selenium-elements/src/main/resources/bordertech-defaults.properties
@@ -1,2 +1,2 @@
-## Use SMART Helper
-bordertech.factory.impl.com.github.bordertech.webfriends.selenium.util.driver.HelperProvider=com.github.bordertech.webfriends.selenium.smart.driver.SmartHelperProvider
+## Use SmartDriver
+bordertech.factory.impl.com.github.bordertech.webfriends.selenium.util.driver.FriendDriver=com.github.bordertech.webfriends.selenium.smart.driver.SmartDriver
